Convert feedback to .vue component
